+++ src/main/java/com/prgrms/ijuju/domain/stock/adv/stockrecord/dto/response/StockRecordResponseDto.java
@@ -16,7 +16,7 @@ public class StockRecordResponseDto {
     private String stockSymbol;
     private TradeType tradeType;
     private BigDecimal price;
-    private int quantity;
+    private double quantity;
     private LocalDateTime tradeDate;
 
     public static StockRecordResponseDto fromEntity(StockRecord stockRecord) {
